requirements.txt
certifi==2019.6.16
chardet==3.0.4
get==2019.4.13
idna==2.8
lxml==4.3.4
post==2019.4.13
public==2019.4.13
PyExecJS==1.5.1
pyjsparser==2.7.1
pytz==2019.1
query-string==2019.4.13
requests==2.22.0
selenium==3.141.0
six==1.12.0
tzlocal==1.5.1
urllib3==1.25.3
